Improved display update function
Solution to plot the display based on the map, the player and the ennemy has been refactored to be way more elegant. Ennemy plot added. Also, it should now allow more flexibility around the edges of the map, to avoid forcing player-centering.

This hasn't been tested yet, and really needs to be.